Add grouped command system
Warning: this commit reveals a bug. When the spinBox is disabled, the
function data of segment model is called with EditRole and writes
something in column 1. This causes an EditCommand to be pushed on the
command stack. The same problem occurs when the new selected segment
have a different combo index than the previous one.
We should change spinBox handling in future commits.