Enhance whole spinbox system
Replace the listview by a tableview.
Embedding spinboxes in this tableview using a custom delegate.
Spares the code from very clumsy and dirty code lines. These ugly lines
were there because of the unicity of the old spinbox. It was a very bad
conception and gave us a lot of headdaches. The whole conception is way
more healthy now.